Moved cleaning game events to separate directory
This should solve #97 and #98
Using a switch case instead of if statements, game event types are exclusive.
Other improvements.
